% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_visits_aggregates.R
\name{aggregate_port_visits}
\alias{aggregate_port_visits}
\title{Title}
\usage{
aggregate_port_visits(
  df,
  squid_aoi,
  min_year = 2017,
  max_year = 2020,
  min_num = 3
)
}
\arguments{
\item{min_num}{}
}
\value{

}
\description{
Title
}
