% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_counts.R
\name{endemic_vessel_counts}
\alias{endemic_vessel_counts}
\title{Count vessels not found in another region}
\usage{
endemic_vessel_counts(df, region)
}
\arguments{
\item{df}{squid vessel positions, dataframe}

\item{region}{squid fishing regions, string}
}
\value{

}
\description{
Count vessels not found in another region
}
