% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_connections_data}
\alias{get_connections_data}
\title{Run queries to load key datasets}
\usage{
get_connections_data(type)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Run queries to load key datasets
}
