% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper-global-squid-package.R
\docType{package}
\name{paper-global-squid-package}
\alias{paper-global-squid}
\alias{paper-global-squid-package}
\title{paper-global-squid: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GlobalFishingWatch/paper-global-squid}
  \item Report bugs at \url{https://github.com/GlobalFishingWatch/paper-global-squid/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
