% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_counts.R
\name{region_connection_count}
\alias{region_connection_count}
\title{Count of connections between regions}
\usage{
region_connection_count(df, combinations)
}
\arguments{
\item{df}{squid vessel postions, dataframe}

\item{combinations}{unique region combinations, list}
}
\value{
dataframe with counts of connections for each region combination
}
\description{
Count of connections between regions
}
