% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_plots}
\alias{save_plots}
\title{Save plots}
\usage{
save_plots(
  plot_list,
  file_name,
  plot_width,
  plot_height,
  units = "mm",
  dpi = 300,
  output_type,
  ...
)
}
\arguments{
\item{dpi}{}
}
\value{

}
\description{
Save plots
}
