% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_visits_aggregates.R
\name{specific_ports_byaoi}
\alias{specific_ports_byaoi}
\title{Title}
\usage{
specific_ports_byaoi(
  df,
  squid_aoi,
  min_year = 2017,
  max_year = 2020,
  min_count = 3
)
}
\arguments{
\item{min_count}{}
}
\value{

}
\description{
Title
}
