% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_counts.R
\name{ssvid_mult_region}
\alias{ssvid_mult_region}
\title{Count vessels one or more regions}
\usage{
ssvid_mult_region(df, num_regions)
}
\arguments{
\item{df}{squid vessel positions, dataframe}

\item{num_regions}{number of regions, integer}
}
\value{

}
\description{
Count vessels one or more regions
}
